Add Creator to Board, Protecting Board Routes

+++ src/controllers/boardController.js
@@ -3,7 +3,7 @@ import Board from '../models/Board';
 
 export const board = async (req, res) => {
   try {
-    const boards = await Board.find({});
+    const boards = await Board.find({}).populate('creator');
     const firstPage = boards.slice(0, 10);
     const pageIndex = [];
     for (let i = 0; i < Math.ceil(boards.length / 10); i++) {
@@ -49,8 +49,11 @@ export const postWriteBoard = async (req, res) => {
   try {
     const newBoard = await Board.create({
       title,
-      content
+      content,
+      creator: req.user.id
     });
+    req.user.boards.push(newBoard.id);
+    req.user.save();
     res.redirect(routes.boardDetail(newBoard.id));
   } catch (error) {
     console.log(error);
@@ -63,7 +66,7 @@ export const boardDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const oneBoard = await Board.findById(id);
+    const oneBoard = await Board.findById(id).populate('creator');
     res.render('boardDetail', { pageTitle: '게시판 내용', oneBoard });
   } catch (error) {
     console.log(error);
@@ -77,7 +80,11 @@ export const getEditBoard = async (req, res) => {
   } = req;
   try {
     const oneBoard = await Board.findById(id);
-    res.render('editBoard', { pageTitle: '게시판 수정', oneBoard });
+    if (board.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editBoard', { pageTitle: '게시판 수정', oneBoard });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.board);
@@ -102,7 +109,12 @@ export const deleteBoard = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Board.findByIdAndRemove(id);
+    const oneBoard = await Board.findById(id);
+    if (oneBoard.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Board.findByIdAndRemove(id);
+    }
     res.redirect(routes.board);
   } catch (error) {
     console.log(error);
